Make BuildDependencies construction a proper function outside of the settings sytem.
