+++ main/src/main/scala/sbt/BuildUtil.scala
@@ -48,6 +48,20 @@ object BuildUtil
 		new BuildUtil(keyIndex, data, root, Load getRootProject units, getp, configs, aggregates)
 	}
 
+	def dependencies(units: Map[URI, LoadedBuildUnit]): BuildDependencies =
+	{
+			import collection.mutable.HashMap
+		val agg = new HashMap[ProjectRef, Seq[ProjectRef]]
+		val cp = new HashMap[ProjectRef, Seq[ClasspathDep[ProjectRef]]]
+		for(lbu <- units.values; rp <- lbu.defined.values)
+		{
+			val ref = ProjectRef(lbu.unit.uri, rp.id)
+			cp(ref) = rp.dependencies
+			agg(ref) = rp.aggregate
+		}
+		BuildDependencies(cp.toMap, agg.toMap)
+	}
+
 	def checkCycles(units: Map[URI, LoadedBuildUnit])
 	{
 		def getRef(pref: ProjectRef) = units(pref.build).defined(pref.project)
